app/requirements.txt
Flask
mysql
mysql-connector
pyopenssl
requests
flask_sqlalchemy
pymysql
